# use endpoint to get both write urls at once