--- a/lightly/openapi_generated/swagger_client/api/__init__.py
+++ b/lightly/openapi_generated/swagger_client/api/__init__.py
@@ -8,6 +8,7 @@ from lightly.openapi_generated.swagger_client.api.embeddings_api import Embeddin
 from lightly.openapi_generated.swagger_client.api.embeddings2d_api import Embeddings2dApi
 from lightly.openapi_generated.swagger_client.api.jobs_api import JobsApi
 from lightly.openapi_generated.swagger_client.api.mappings_api import MappingsApi
+from lightly.openapi_generated.swagger_client.api.other_api import OtherApi
 from lightly.openapi_generated.swagger_client.api.quota_api import QuotaApi
 from lightly.openapi_generated.swagger_client.api.samples_api import SamplesApi
 from lightly.openapi_generated.swagger_client.api.samplings_api import SamplingsApi