… lightly/openapi_generated/swagger_client/api/other_api.py
… lightly/openapi_generated/swagger_client/models/sample_write_urls.py