closes #420

### Other change not related to the issue:
The lightly-crop feature is now more verbose and uses tqdm:
https://github.com/lightly-ai/lightly/blob/6e2378225dea4bdb85d2556d305a65d3865ce995/lightly/cli/crop_cli.py#L50-L51


## Description
This uses the new endpoint to get both write urls (thumbnail and full image) at once. It gives a slight performance boost (about 10%) both with 0 and 12 workers.

## Performance measurement with num_workers=0
- 241 images from clothing dataset
#### with this PR: 136s
<img width="603" alt="image" src="https://user-images.githubusercontent.com/20324507/123610717-227b3980-d801-11eb-8d21-ac6af270b967.png">

#### before this PR: 156s
<img width="518" alt="image" src="https://user-images.githubusercontent.com/20324507/123611631-1643ac00-d802-11eb-99ce-ea9bd5f6240b.png">

## Performance measurement with num_workers=12
- 241 images from clothing dataset
#### with this PR: 14.2s
<img width="645" alt="image" src="https://user-images.githubusercontent.com/20324507/123612900-40499e00-d803-11eb-97ad-f1d37db87151.png">


#### before this PR: 15.7s
<img width="509" alt="image" src="https://user-images.githubusercontent.com/20324507/123611997-733f6200-d802-11eb-9215-ec1e433d9ba3.png">